+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.18.0
+------------------------------
+*April 5, 2019*
+
+### Added
+- Tests ensure urls point to a page returning a 200 status code
+- Manual test readme docs
+
+### Changed
+- Footer links (about us, cuisine, location)
+
+
 v1.17.0
 ------------------------------
 *March 12, 2019*